Improve admin functionality

+++ app/src/main/java/com/pr0gramm/app/api/pr0gramm/Api.java
@@ -147,7 +147,7 @@ public interface Api {
             @Field("customReason") String customReason,
             @Field("notifyUser") String notifyUser,
             @Field("banUser") String banUser,
-            @Field("days") Integer days);
+            @Field("days") Float days);
 
     @FormUrlEncoded
     @POST("api/profile/follow")
